+++ packages/components-props-analyzer/src/ComponentsList.ts
@@ -94,6 +94,15 @@ const components: Component[] = [
         packageName: '@coveord/plasma-mantine',
     },
     {
+        name: 'Badge',
+        packageName: '@coveord/plasma-mantine',
+    },
+    {
+        name: 'InfoToken',
+        packageName: '@coveord/plasma-mantine',
+        propsType: 'InfoTokenProps',
+    },
+    {
         name: 'EllipsisText',
         packageName: '@coveord/plasma-mantine',
         propsType: 'EllipsisTextProps',
@@ -112,6 +121,11 @@ const components: Component[] = [
         name: 'DateRangePickerInlineCalendar',
         packageName: '@coveord/plasma-mantine',
         propsType: 'DateRangePickerInlineCalendarProps',
+    },
+    {
+        name: 'DateTimeRangePicker',
+        packageName: '@coveord/plasma-mantine',
+        propsType: 'DateTimeRangePickerProps',
     },
 ];