+++ packages/mantine/src/components/copyToClipboard/CopyToClipboard.tsx
@@ -1,8 +1,9 @@
-import {CheckSize16Px, CopySize16Px} from '@coveord/plasma-react-icons';
-import {ActionIcon, CopyButton, MantineColor, TextInput, Tooltip} from '@mantine/core';
-import {FunctionComponent} from 'react';
+import {IconClipboardCheck, IconClipboardText} from '@coveord/plasma-react-icons';
+import {ActionIconProps, CopyButton, MantineColor, TextInput, Tooltip} from '@mantine/core';
+import {FunctionComponent, MouseEventHandler} from 'react';
+import {ActionIcon} from '../action-icon';
 
-export interface CopyToClipboardProps {
+export interface CopyToClipboardProps extends ActionIconProps {
     /**
      * The value to be copied to the clipboard.
      */
@@ -16,29 +17,49 @@ export interface CopyToClipboardProps {
     /**
      * Called each time the value is copied to the clipboard
      */
-    onCopy?: () => void;
+    onCopy?: MouseEventHandler<HTMLButtonElement>;
     /**
      * The color of the icon when idle
      *
      * @default 'gray'
      */
     color?: MantineColor | (string & {});
+    /**
+     * The text displayed when hovering over the button
+     *
+     * @default 'Copy to clipboard'
+     */
+    tooltipLabelCopy?: string;
+    /**
+     * The text displayed when the value is copied to the clipboard
+     *
+     * @default 'Copied'
+     */
+    tooltipLabelCopied?: string;
 }
 
-const CopyToClipboardButton: FunctionComponent<Omit<CopyToClipboardProps, 'withLabel'>> = ({value, onCopy, color}) => (
+const CopyToClipboardButton: FunctionComponent<Omit<CopyToClipboardProps, 'withLabel'>> = ({
+    value,
+    onCopy,
+    color,
+    tooltipLabelCopy = 'Copy to clipboard',
+    tooltipLabelCopied = 'Copied',
+    ...others
+}) => (
     <CopyButton value={value} timeout={2000}>
         {({copied, copy}) => (
-            <Tooltip label={copied ? 'Copied' : 'Copy'}>
-                <ActionIcon
-                    variant="subtle"
+            <Tooltip label={copied ? tooltipLabelCopied : tooltipLabelCopy}>
+                <ActionIcon.Quaternary
+                    aria-label={tooltipLabelCopy}
                     color={copied ? 'success' : color}
-                    onClick={() => {
+                    onClick={(clickEvent) => {
                         copy();
-                        onCopy?.();
+                        onCopy?.(clickEvent);
                     }}
+                    {...others}
                 >
-                    {copied ? <CheckSize16Px height={16} /> : <CopySize16Px height={16} />}
-                </ActionIcon>
+                    {copied ? <IconClipboardCheck size={20} /> : <IconClipboardText size={20} />}
+                </ActionIcon.Quaternary>
             </Tooltip>
         )}
     </CopyButton>
@@ -50,7 +71,7 @@ export const CopyToClipboard: FunctionComponent<CopyToClipboardProps> = ({withLa
             value={others.value}
             readOnly
             autoComplete="off"
-            rightSection={<CopyToClipboardButton color="action" {...others} />}
+            rightSection={<CopyToClipboardButton {...others} />}
         />
     ) : (
         <CopyToClipboardButton {...others} />