+++ packages/mantine/src/index.ts
@@ -5,6 +5,7 @@ import {type PlasmaColors} from './theme/PlasmaColors';
 export * from '@mantine/carousel';
 export * from '@mantine/core';
 export {Pagination} from '@mantine/core';
+export {type DatesRangeValue} from '@mantine/dates';
 export * from '@mantine/form';
 export * from '@mantine/hooks';
 export * from '@mantine/notifications';
@@ -15,6 +16,7 @@ export {noop};
 
 // explicitly overriding mantine components
 export {
+    Accordion,
     ActionIcon,
     BrowserPreview,
     Button,