+++ packages/mantine/src/theme/Theme.tsx
@@ -1,19 +1,19 @@
 import {
     ArrowHeadLeftSize16Px,
     ArrowHeadRightSize16Px,
-    CheckSize16Px,
     CrossSize16Px,
     FilterSize16Px,
-    InfoSize16Px,
+    IconCheck,
+    IconSlash,
     InfoSize24Px,
 } from '@coveord/plasma-react-icons';
 import {color} from '@coveord/plasma-tokens';
 import {
     ActionIcon,
     Alert,
-    Anchor,
-    AppShellNavbar,
+    AppShell,
     Badge,
+    Breadcrumbs,
     Button,
     Checkbox,
     CloseButton,
@@ -28,7 +28,6 @@ import {
     List,
     Loader,
     MantineThemeOverride,
-    MenuItem,
     Modal,
     MultiSelect,
     NavLink,
@@ -42,27 +41,25 @@ import {
     Skeleton,
     Stepper,
     Switch,
+    Table,
     Tabs,
     Text,
-    TextInput,
     Tooltip,
 } from '@mantine/core';
-import {DatePicker} from '@mantine/dates';
+import {CheckboxIcon, CircleLoader, InfoToken} from '../components';
 import ActionIconClasses from '../styles/ActionIcon.module.css';
 import AlertClasses from '../styles/Alert.module.css';
-import AnchorClasses from '../styles/Anchor.module.css';
-import AppShellNavBarClasses from '../styles/AppShellNavBar.module.css';
 import BadgeClasses from '../styles/Badge.module.css';
+import BreadcrumbsClasses from '../styles/Breadcrumbs.module.css';
 import ButtonClasses from '../styles/Button.module.css';
 import CheckboxClasses from '../styles/Checkbox.module.css';
+import CheckboxIndicatorClasses from '../styles/CheckboxIndicator.module.css';
 import ComboboxClasses from '../styles/Combobox.module.css';
-import DatePickerClasses from '../styles/DatePicker.module.css';
 import InputClasses from '../styles/Input.module.css';
 import InputWrapperClasses from '../styles/InputWrapper.module.css';
 import ListClasses from '../styles/List.module.css';
 import ModalClasses from '../styles/Modal.module.css';
 import NavLinkClasses from '../styles/NavLink.module.css';
-import NotificationClasses from '../styles/Notification.module.css';
 import PaginationClasses from '../styles/Pagination.module.css';
 import RadioClasses from '../styles/Radio.module.css';
 import ReadOnlyInputClasses from '../styles/ReadOnlyInput.module.css';
@@ -72,18 +69,17 @@ import SegmentedControlClasses from '../styles/SegmentedControl.module.css';
 import SelectClasses from '../styles/Select.module.css';
 import SkeletonClasses from '../styles/Skeleton.module.css';
 import StepperClasses from '../styles/Stepper.module.css';
+import TableClasses from '../styles/Table.module.css';
 import TabsClasses from '../styles/Tabs.module.css';
 import TextClasses from '../styles/Text.module.css';
 import TooltipClasses from '../styles/Tooltip.module.css';
-import {NotificationVars} from '../vars/Notification.vars';
-import {TextVars} from '../vars/Text.vars';
 import {PlasmaColors} from './PlasmaColors';
 
 export const plasmaTheme: MantineThemeOverride = createTheme({
     // These are overrides over https://github.com/mantinedev/mantine/blob/master/packages/%40mantine/core/src/core/MantineProvider/default-theme.ts
     fontFamily: 'canada-type-gibson, sans-serif',
     black: color.primary.gray[9],
-    defaultRadius: 8,
+    defaultRadius: 'md',
     lineHeights: {md: '1.5'},
     spacing: {
         xxs: '4px',
@@ -93,59 +89,89 @@ export const plasmaTheme: MantineThemeOverride = createTheme({
         lg: '32px',
         xl: '40px',
     },
-    primaryColor: 'action',
+    radius: {
+        none: '0px',
+        xs: '2px',
+        sm: '4px',
+        md: '8px',
+        lg: '16px',
+        xl: '24px',
+        xxl: '32px',
+    },
+    primaryColor: 'blue',
     headings: {
         fontFamily: 'canada-type-gibson, sans-serif',
         fontWeight: '500',
         sizes: {
-            h1: {fontSize: '48px', lineHeight: '1.5', fontWeight: '300'},
-            h2: {fontSize: '32px', lineHeight: '1.5', fontWeight: '500'},
-            h3: {fontSize: '24px', lineHeight: '1.5', fontWeight: '500'},
-            h4: {fontSize: '18px', lineHeight: '1.5', fontWeight: '300'},
-            h5: {fontSize: '14px', lineHeight: '1.5', fontWeight: '500'},
-            h6: {fontSize: '12px', lineHeight: '1.5', fontWeight: '500'},
+            h1: {fontSize: '40px', lineHeight: '1.2', fontWeight: '500'},
+            h2: {fontSize: '32px', lineHeight: '1.25', fontWeight: '500'},
+            h3: {fontSize: '24px', lineHeight: '1.33', fontWeight: '500'},
+            h4: {fontSize: '18px', lineHeight: '1.2', fontWeight: '400'},
+            h5: {fontSize: '16px', lineHeight: '1.25', fontWeight: '500'},
+            h6: {fontSize: '12px', lineHeight: '1.33', fontWeight: '500'},
         },
     },
-    shadows: {
-        xs: '0px 1px 0px rgba(4, 8, 31, 0.08)',
-        sm: '0px 2px 4px rgba(4, 8, 31, 0.12)',
-        md: '0px 4px 8px rgba(4, 8, 31, 0.08)',
-        lg: '0px 8px 16px rgba(7, 12, 41, 0.06)',
-        xl: '0px 16px 24px rgba(4, 8, 31, 0.06)',
+    fontSizes: {
+        xs: '12px',
+        sm: '14px',
+        md: '16px',
+        lg: '18px',
+        xl: '20px',
     },
+    shadows: {
+        xs: '0px 1px 2px 0px rgba(0, 0, 0, 0.10), 0px 1px 3px 0px rgba(0, 0, 0, 0.05), 0px -0.5px 1px 0px rgba(0, 0, 0, 0.02)',
+        sm: '0px 7px 7px -5px rgba(0, 0, 0, 0.04), 0px 10px 15px -5px rgba(0, 0, 0, 0.10), 0px 1px 3px 0px rgba(0, 0, 0, 0.05)',
+        md: '0px 10px 10px -5px rgba(0, 0, 0, 0.04), 0px 20px 25px -5px rgba(0, 0, 0, 0.10), 0px 1px 3px 0px rgba(0, 0, 0, 0.05)',
+        lg: '0px 12px 12px -7px rgba(0, 0, 0, 0.04), 0px 28px 23px -7px rgba(0, 0, 0, 0.10), 0px 1px 3px 0px rgba(0, 0, 0, 0.05)',
+        xl: '0px 17px 17px -7px rgba(0, 0, 0, 0.04), 0px 36px 28px -7px rgba(0, 0, 0, 0.10), 0px 1px 3px 0px rgba(0, 0, 0, 0.05)',
+    },
+    primaryShade: 5,
     colors: PlasmaColors,
     components: {
         ActionIcon: ActionIcon.extend({
-            classNames: {root: ActionIconClasses.root},
+            defaultProps: {
+                size: 24,
+                radius: 'sm',
+            },
+            classNames: ActionIconClasses,
+        }),
+        AppShell: AppShell.extend({
+            vars: (theme) =>
+                ({
+                    root: {'--app-shell-border-color': theme.colors.gray[2]},
+                }) as any,
         }),
         Alert: Alert.extend({
             defaultProps: {
-                icon: <InfoSize16Px height={16} />,
-                color: 'navy',
+                icon: <InfoToken variant="advice" />,
+                p: 'sm',
             },
             classNames: AlertClasses,
         }),
-        Anchor: Anchor.extend({
-            defaultProps: {
-                color: 'action.6',
-            },
-            classNames: {root: AnchorClasses.root},
-        }),
-        AppShellNavbar: AppShellNavbar.extend({
-            classNames: {navbar: AppShellNavBarClasses.navbar},
-        }),
         Badge: Badge.extend({
-            classNames: {root: BadgeClasses.root},
+            classNames: BadgeClasses,
             defaultProps: {
                 variant: 'light',
+                tt: 'none',
+            },
+        }),
+        Breadcrumbs: Breadcrumbs.extend({
+            classNames: BreadcrumbsClasses,
+            defaultProps: {
+                separator: <IconSlash size={16} color="var(--mantine-color-dimmed)" />,
+                separatorMargin: 'xxs',
             },
         }),
         Button: Button.extend({
+            defaultProps: {
+                size: 'sm',
+            },
             classNames: ButtonClasses,
         }),
         Checkbox: Checkbox.extend({
             defaultProps: {
                 radius: 'sm',
+                icon: CheckboxIcon,
             },
             classNames: (theme, props) => {
                 if (props.readOnly && !props.disabled) {
@@ -153,17 +179,13 @@ export const plasmaTheme: MantineThemeOverride = createTheme({
                 }
                 return CheckboxClasses;
             },
-            vars: (theme, props) => {
-                if (props.readOnly && !props.disabled) {
-                    return {
-                        root: {
-                            '--checkbox-icon-color': theme.colors.gray[7],
-                            '--checkbox-color': theme.colors.gray[2],
-                        },
-                    };
-                }
-                return {root: {}};
+        }),
+        CheckboxIndicator: Checkbox.Indicator.extend({
+            defaultProps: {
+                radius: 'sm',
+                icon: CheckboxIcon,
             },
+            classNames: CheckboxIndicatorClasses,
         }),
         CloseButton: CloseButton.extend({
             defaultProps: {
@@ -177,7 +199,10 @@ export const plasmaTheme: MantineThemeOverride = createTheme({
             },
         }),
         Combobox: Combobox.extend({
-            classNames: {option: SelectClasses.option, search: ComboboxClasses.search},
+            classNames: ComboboxClasses,
+            defaultProps: {
+                middlewares: {inline: true},
+            },
         }),
         ComboboxSearch: ComboboxSearch.extend({
             defaultProps: {
@@ -185,15 +210,15 @@ export const plasmaTheme: MantineThemeOverride = createTheme({
                 rightSection: <FilterSize16Px height={16} color="gray.5" />,
             },
         }),
-        DatePicker: DatePicker.extend({
-            classNames: {monthCell: DatePickerClasses.monthCell},
-        }),
         Divider: Divider.extend({
             defaultProps: {
-                color: 'gray.3',
+                color: 'var(--mantine-color-default-border)',
             },
         }),
         Input: Input.extend({
+            defaultProps: {
+                size: 'sm',
+            },
             classNames: (_theme, props) => {
                 const anyProps = props as any;
                 if (anyProps.readOnly && !props.disabled && !['Select'].includes(anyProps.__staticSelector)) {
@@ -203,17 +228,24 @@ export const plasmaTheme: MantineThemeOverride = createTheme({
             },
         }),
         InputWrapper: InputWrapper.extend({
+            defaultProps: {
+                size: 'sm',
+            },
             classNames: InputWrapperClasses,
             vars: (theme, props) => {
                 const anyProps = props as any;
                 if (anyProps.disabled || (anyProps.readOnly && !['Select'].includes(anyProps.__staticSelector))) {
                     return {
+                        root: {
+                            '--input-margin-top': props?.label || props?.description ? theme.spacing.xxs : undefined,
+                        },
                         label: {'--input-asterisk-color': theme.colors.red[2]},
                         error: {},
                         description: {},
                     };
                 }
                 return {
+                    root: {'--input-margin-top': props?.label || props?.description ? theme.spacing.xxs : undefined},
                     label: {},
                     error: {},
                     description: {},
@@ -222,26 +254,36 @@ export const plasmaTheme: MantineThemeOverride = createTheme({
         }),
         Loader: Loader.extend({
             defaultProps: {
-                type: 'dots',
-                color: 'action',
+                loaders: {...Loader.defaultLoaders, circle: CircleLoader},
+                type: 'circle',
                 role: 'presentation',
             },
         }),
         List: List.extend({
-            classNames: {root: ListClasses.root},
-        }),
-        MenuItem: MenuItem.extend({
-            defaultProps: {
-                fw: 300,
-            },
+            classNames: ListClasses,
         }),
         Modal: Modal.extend({
             classNames: ModalClasses,
+            vars: () => {
+                const sizes = {
+                    '--modal-size-xs': '432px',
+                    '--modal-size-sm': '664px',
+                    '--modal-size-md': '896px',
+                    '--modal-size-lg': '1120px',
+                    '--modal-size-xl': '88%',
+                } as any;
+                return {
+                    root: {
+                        ...sizes,
+                    },
+                };
+            },
         }),
         ModalOverlay: Modal.Overlay.extend({
             defaultProps: {
-                color: color.primary.navy[9],
-                backgroundOpacity: 0.9,
+                color: PlasmaColors.indigo[8],
+                backgroundOpacity: 0.7,
+                blur: 2,
             },
         }),
         ModalRoot: Modal.Root.extend({
@@ -254,18 +296,10 @@ export const plasmaTheme: MantineThemeOverride = createTheme({
         Notification: Notification.extend({
             defaultProps: {
                 icon: <InfoSize24Px height={24} />,
-                color: 'info',
             },
-            classNames: {
-                root: NotificationClasses.root,
-                icon: NotificationClasses.icon,
-                closeButton: NotificationClasses.closeButton,
-            },
-            vars: NotificationVars,
         }),
         Pagination: Pagination.extend({
             classNames: PaginationClasses,
-            vars: () => ({root: {'--pagination-control-fz': 'var(--mantine-font-size-sm)'}}),
             defaultProps: {
                 nextIcon: ArrowHeadRightSize16Px,
                 previousIcon: ArrowHeadLeftSize16Px,
@@ -275,6 +309,7 @@ export const plasmaTheme: MantineThemeOverride = createTheme({
             defaultProps: {
                 shadow: 'md',
                 withArrow: true,
+                middlewares: {inline: true},
             },
         }),
         Radio: Radio.extend({
@@ -297,31 +332,24 @@ export const plasmaTheme: MantineThemeOverride = createTheme({
             },
         }),
         ScrollArea: ScrollArea.extend({
-            classNames: {viewport: ScrollAreaClasses.viewport},
+            classNames: ScrollAreaClasses,
         }),
         SegmentedControl: SegmentedControl.extend({
             classNames: SegmentedControlClasses,
         }),
         Select: Select.extend({
-            defaultProps: {withCheckIcon: false, allowDeselect: false},
-            classNames: {input: SelectClasses.input, option: SelectClasses.option},
+            defaultProps: {allowDeselect: false, withCheckIcon: false},
+            classNames: SelectClasses,
         }),
         Skeleton: Skeleton.extend({
-            classNames: {root: SkeletonClasses.root},
+            classNames: SkeletonClasses,
         }),
         Stepper: Stepper.extend({
             defaultProps: {
                 size: 'xs',
-                completedIcon: <CheckSize16Px />,
+                completedIcon: <IconCheck size={16} />,
             },
-            classNames: {
-                step: StepperClasses.step,
-                stepIcon: StepperClasses.stepIcon,
-                stepCompletedIcon: StepperClasses.stepCompletedIcon,
-                stepDescription: StepperClasses.stepDescription,
-                separator: StepperClasses.separator,
-                verticalSeparator: StepperClasses.verticalSeparator,
-            },
+            classNames: StepperClasses,
         }),
         Switch: Switch.extend({
             classNames: (theme, props) => {
@@ -345,27 +373,29 @@ export const plasmaTheme: MantineThemeOverride = createTheme({
                 }
                 return {root: {}, track: {}, thumb: {}};
             },
+            defaultProps: {
+                withThumbIndicator: false,
+            },
+        }),
+        Table: Table.extend({
+            classNames: TableClasses,
+            vars: () => ({
+                table: {
+                    '--table-border-color': 'var(--mantine-color-default-border)',
+                },
+            }),
         }),
         Tabs: Tabs.extend({
             classNames: TabsClasses,
         }),
-        TabsTab: Tabs.Tab.extend({
-            defaultProps: {
-                px: 'sm',
-            },
-        }),
         Text: Text.extend({
-            classNames: TextClasses,
-            vars: TextVars,
-        }),
-        TextInput: TextInput.extend({
             defaultProps: {
-                radius: 8,
+                size: 'sm',
             },
+            classNames: TextClasses,
         }),
         Tooltip: Tooltip.extend({
             defaultProps: {
-                color: 'navy',
                 maw: 300,
                 multiline: true,
                 withArrow: true,