+++ packages/website/src/examples/feedback/alert/AlertAdvice.demo.tsx
@@ -1,0 +1,9 @@
+import {AlertAdvice} from '@coveord/plasma-mantine';
+
+const Demo = () => (
+    <AlertAdvice title="Advice" withCloseButton>
+        Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore
+        magna aliqua.
+    </AlertAdvice>
+);
+export default Demo;