+++ packages/website/src/examples/feedback/alert/AlertCritical.demo.tsx
@@ -1,10 +1,9 @@
-import {Alert} from '@coveord/plasma-mantine';
-import {CriticalSize16Px} from '@coveord/plasma-react-icons';
+import {AlertCritical} from '@coveord/plasma-mantine';
 
 const Demo = () => (
-    <Alert title="Attention!" withCloseButton color="critical" icon={<CriticalSize16Px height={16} />}>
+    <AlertCritical title="Bummer!" withCloseButton>
         Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore
         magna aliqua.
-    </Alert>
+    </AlertCritical>
 );
 export default Demo;